requirements.txt
requests
beautifulSoup4
pillow
selenium
webdriver-manager
pandas
opencv-python
pytesseract
colorthief